rename ProofCombinators module
