+++ benchmarks/popl18/nople/neg/Append.hs
@@ -40,7 +40,8 @@ prop_append_neutral N
 prop_append_neutral (C x xs)
   = toProof $ 
        append (C x xs) N === C x (append xs N)
-                         ==? C x xs             ? prop_append_neutral xs  
+                           ? prop_append_neutral xs  
+                         === C x xs             
 
 {-@ prop_assoc :: xs:L a -> ys:L a -> zs:L a
                -> {v:Proof | append (append xs ys) zs /= append xs (append ys zs) } @-}
@@ -55,7 +56,8 @@ prop_assoc (C x xs) ys zs
       append (append (C x xs) ys) zs
         === append (C x (append xs ys)) zs
         === C x (append (append xs ys) zs)
-        ==? C x (append xs (append ys zs))  ? prop_assoc xs ys zs 
+          ? prop_assoc xs ys zs 
+        === C x (append xs (append ys zs)) 
         === append (C x xs) (append ys zs)
 
 
@@ -75,8 +77,8 @@ prop_map_append f (C x xs) ys
        map f (append (C x xs) ys)
          === map f (C x (append xs ys))
          === C (f x) (map f (append xs ys))
-         ==? C (f x) (append (map f xs) (map f ys)) 
              ? prop_map_append f xs ys
+         === C (f x) (append (map f xs) (map f ys)) 
          === append (C (f x) (map f xs)) (map f ys)
          === append (map f (C x xs)) (map f ys)
 
@@ -96,11 +98,9 @@ prop_concatMap f (C x xs)
       concatt (map f (C x xs))
         === concatt (C (f x) (map f xs))
         === append (f x) (concatt (map f xs))
-        ==? append (f x) (concatMap f xs)       
             ? prop_concatMap f xs 
+        === append (f x) (concatMap f xs)       
         === concatMap f (C x xs)
-
-
 
 {-@ data L [llen] @-}
 data L a = N | C a (L a)
