+++ benchmarks/popl18/nople/pos/MapFusion.hs
@@ -30,7 +30,8 @@ map_fusion f g N
 map_fusion f g (C x xs)
   = map (compose f g) (C x xs)
        === C ((compose f g) x) (map (compose f g) xs)
-       ==? C ((compose f g) x) ((compose (map f) (map g)) xs) ? map_fusion f g xs
+         ? map_fusion f g xs
+       === C ((compose f g) x) ((compose (map f) (map g)) xs) 
        === C ((compose f g) x) (map f (map g xs))
        === C (f (g x)) (map f (map g xs))
        === map f (C (g x) (map g xs))
