+++ benchmarks/popl18/nople/pos/Overview.hs
@@ -46,7 +46,8 @@ safe () =
 {-@ safe' :: () ->  { fib 3 == 2 } @-}
 safe' () 
   =   fib 3 
-  ==? fib 2 + fib 1 ? safe ()
+    ? safe ()
+  === fib 2 + fib 1 
   === 2
   *** QED
 
@@ -71,9 +72,9 @@ fib_incr n
    | otherwise
    = fib n
    === fib (n-1) + fib (n-2)
-   =<=? fib n     + fib (n-2)
-           ? fib_incr (n-1)
-   =<=? fib n     + fib (n-1)
-           ? fib_incr (n-2)
-   =<=? fib (n+1)
+     ? fib_incr (n-1)
+   =<= fib n     + fib (n-2)
+     ? fib_incr (n-2)
+   =<= fib n     + fib (n-1)
+   =<= fib (n+1)
    *** QED
