+++ benchmarks/popl18/nople/pos/Peano.hs
@@ -49,7 +49,8 @@ zeroR Z
 zeroR (S n)
   =   plus (S n) Z
   === S (plus n Z)
-  ==? S n                      ? zeroR n
+    ? zeroR n
+  === S n
   *** QED
 
 {-@ plusSuccR :: n:Peano -> m:Peano -> { plus n (S m) = S (plus n m) } @-}
@@ -62,21 +63,25 @@ plusSuccR Z m
 plusSuccR (S n) m
   =   plus (S n) (S m)
   === S (plus n (S m))
-  ==? S (S (plus n m))        ? plusSuccR n m
+      ? plusSuccR n m
+  === S (S (plus n m)) 
   === S (plus (S n) m)
   *** QED
 
 {-@ plusComm :: a:_ -> b:_  -> {plus a b == plus b a} @-}
 plusComm Z b
   =   plus Z b
-  ==? plus b Z                ? zeroR b
+      ? zeroR b
+  === plus b Z 
   *** QED
 
 plusComm (S a) b
   =   plus (S a) b
   === S (plus a b)
-  ==? S (plus b a)            ? plusComm a b
-  ==? plus b (S a)            ? plusSuccR b a
+    ? plusComm a b
+  === S (plus b a) 
+    ? plusSuccR b a
+  === plus b (S a)      
   *** QED
 
 {-@ plusAssoc :: a:_ -> b:_ -> c:_ -> {plus (plus a b) c == plus a (plus b c) } @-}
@@ -90,6 +95,7 @@ plusAssoc (S a) b c
   =   plus (plus (S a) b) c
   === plus (S (plus a b)) c
   === S (plus (plus a b) c)
-  ==? S (plus a (plus b c))   ? plusAssoc a b c
+    ? plusAssoc a b c
+  === S (plus a (plus b c)) 
   === plus (S a) (plus b c)
   *** QED
